Updated Directory Structure
