+++ Array/kadane_algorithm.cpp
