+++ Graphs/Kruskalsalgorithm.cpp
