+++ Leetcode/42TrappingRainWater.cpp
