+++ Math/armstrong.cpp
